Added new logging and better command response
When an everyone or here mention is seen, it's logged in the
notification channel.
To determine if a user sends "why, y tho, etc.", I send the message to
tools.messageIsWhyCmd which figures out if it matches one of the
available inputs then returns true or false if it matches.