Added kill command and command use tracking
When a command is used, an entry in the local MySQL database is increased by 1. This way I can keep track of the most popular commands.

Also added a kill command that uses a new method for displaying images. This is the way I'd like to do all future commands so this is the testing ground.