Base marriage supported added
Commands such as `+marry @Alcha#2621` and `+marry @Alcha#2621 list` now work but still need some testing.

Standard divorces (as they were in the previous version of Tron) will be the initial goal, with the hope to add a "forced" divorce option in the event the other individual is causing problems or prolonging the inevitable.