Preliminary fix for marry command
I fixed the marriage command so that it displays the proper users and marriage lists. Before, it was showing the information for previous commands and was unsafe in terms of which id was used.

New cmd structure ensures the db connection is created on each use. This is a dirty way of doing it, but it'll get things working for now.