Added LeaveRole command
Now users are able to join/leave roles as they please. The only requirement to join a role is that it be listed in the servers list of available roles.

There is no requirement to try and delete a role so long as you provide a role or role id to the removeRole method.