Birthdays command added
Tron now accepts birthdays as input in order to store a users birthday. When that date comes around, using the BabyTron repo with node-ipc, Tron receives that users id and they are sent a birthday message wishing them a happy birthday.

Later, the birthday notification will also be sent to servers who have the option enabled.