Added some command logging to help locate cause of crash
There's been a crash recently where Tron lacks permission to perform an action. Unfortunately, I can't locate where the error is but I am able to catch the error in a log with a timestamp.

Now, I log each command before it's executed with the timestamp of when it was executed. This should help out tremendously.