Added support for displaying types of emoji with emote command
A user can now use `+emotes still` to only display the still/non-animated emoji or `+emotes animated` to only display the animated emoji.

It also accepts an all argument in lieu of animated or still which means to display both animated and non-animated emoji. This is the default behavior if no arguments are provided to the command.