Stashing the current state of Polls; Survaid needs work to continue
To continue with the poll command, I have to fix a bug in Survaid as well as refactor it in order to support the arguments I now deem required, whereas before I was getting feature crazy and kept adding new arguments. I need to pare it down and get it ready for Tron.